refactor(links): move test
